[Snippets] Added SubgraphPass and SubgraphManager support (PR#19951)
